+++ lib/src/rules_factory.dart
@@ -1,5 +1,6 @@
 import 'rules/avoid_preserve_whitespace_false.dart';
 import 'rules/base_rule.dart';
+import 'rules/binary_expression_operand_order_rule.dart';
 import 'rules/double_literal_format_rule.dart';
 import 'rules/newline_before_return.dart';
 import 'rules/no_boolean_literal_compare_rule.dart';
@@ -15,6 +16,7 @@ const _implementedRules = {
   NoEmptyBlockRule(),
   NoMagicNumberRule(),
   PreferTrailingCommaForCollectionRule(),
+  BinaryExpressionOperandOrderRule(),
 };
 
 Iterable<BaseRule> get allRules => _implementedRules;