+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Unreleased
 - Drop dependency on resource
+- Add binary-expression-operand-order rule
 
 # 1.6.0
 - Add experimental static code diagnostics prefer-trailing-comma-for-collection, no-magic-number rule